Rename files to better fit Turso tutorial
